requirements.txt
zhipuai>=1.0.0
python-dotenv>=1.0.0
pypiwin32>=223; platform_system == "Windows"
